% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{check_attributes_biomart}
\alias{check_attributes_biomart}
\title{Check available attributes (key types) for selected organism}
\usage{
check_attributes_biomart(organism, host = "www.ensembl.org")
}
\arguments{
\item{organism}{Organism to check attributes from. Can be checked using check_organisms_biomart()}

\item{host}{Host server for Ensembl database, closest server will automatically
selected. However if there is an error, it can be set to the following:}
}
\value{
List of attributes()
}
\description{
Check available attributes (key types) for selected organism
}
\examples{
check_attributes_biomart("ggallus", host = "asia.ensembl.org")
}
