% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{check_organisms_biomart}
\alias{check_organisms_biomart}
\title{Check organism biomart (Ensembl database)}
\usage{
check_organisms_biomart()
}
\value{
List of organism that are available on the Ensembl database.
}
\description{
Check organism biomart (Ensembl database)
}
\examples{
check_organisms_biomart()
}
