% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{convert_biomart}
\alias{convert_biomart}
\title{Convert ID using biomaRt (ensembl database)}
\usage{
convert_biomart(
  genelist,
  organism = "hsapiens",
  supplied_id = "ensembl_gene_id",
  attribs = c("ensembl_gene_id", "external_gene_name", "description",
    "mmusculus_homolog_ensembl_gene"),
  host = "www.ensembl.org",
  unique_rows = FALSE,
  replace_extgenename_with_ensembl = T,
  remove_geneid_version = T
)
}
\arguments{
\item{genelist}{List of gene IDs in the form of character vector.}

\item{organism}{Organism of the input gene ID (can be checked using check_organisms_biomart()).}

\item{supplied_id}{The key of the gene ID supplied (e.g. "ensembl_gene_id" or
"external_gene_name"). Can be checked using check_attributes_biomart().}

\item{attribs}{Gene ID to convert the input ID to. Can be checked using 
check_attributes_biomart().}

\item{host}{Host server for Ensembl database, closest server will automatically
selected. However if there is an error, it can be set to the following:
"uswest.ensembl.org", "asia.ensembl.org", "useast.ensembl.org".}

\item{unique_rows}{Removes duplicated rows.}

\item{replace_extgenename_with_ensembl}{When converting ensemblid and there 
is no corresponding gene name, replace with ensembl gene id.}

\item{remove_geneid_version}{Remove gene id version from ensembl ID (e.g. ENSGALG00000025996.4 to ENSGALG00000025996).}
}
\value{
data frame consisting of gene ids.
}
\description{
Convert ID using biomaRt (ensembl database)
}
\examples{
convert_biomart(c("ENSGALG00000025996", "ENSGALG00000017312.5"), organism = "ggallus")
}
