% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tcga.R
\name{convert_tcga}
\alias{convert_tcga}
\title{Convert TCGA Identifiers}
\usage{
convert_tcga(x, from = "sample_id", to = "submitter_id", multiple = FALSE)
}
\arguments{
\item{x}{A character vector to convert.}

\item{from}{Which identifier type to be converted. One of case_id, aliquot_ids, submitter_aliquot_ids, submitter_id, sample_id.}

\item{to}{Identifier type convert to. Same as parameter \code{from}.}

\item{multiple}{if \code{TRUE}, return a \code{data.table} instead of a
string vector, so multiple identifier mappings can be kept.}
}
\value{
A character vector.
}
\description{
Run \code{data("tcga")} to see detail database for conversion.
}
\examples{
\donttest{
x <- convert_tcga("TCGA-02-0001-10")
x
}
\dontrun{
convert_tcga("TCGA-02-0001-10A-01W-0188-10")
}
}
