% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\docType{data}
\name{icgc}
\alias{icgc}
\title{ICGC Sample Identifiers}
\format{
A data frame with 155874 rows and 6 variables.
}
\source{
\url{https://dcc.icgc.org/repositories}
}
\description{
ICGC Sample Identifiers
}
\examples{
\donttest{
load_data("icgc")
}
}
