% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\docType{data}
\name{pcawg_full}
\alias{pcawg_full}
\title{PCAWG Full Sample Identifiers}
\format{
A data frame with 7255 rows and 8 variables.
}
\source{
\url{https://dcc.icgc.org/releases/PCAWG}
}
\usage{
pcawg_full
}
\description{
PCAWG Full Sample Identifiers
}
\examples{
data("pcawg_full")
}
\keyword{datasets}
