% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\docType{data}
\name{pcawg_simple}
\alias{pcawg_simple}
\title{PCAWG Mutation Related Simplified Sample Identifiers}
\format{
A data frame with 2583 rows and 12 variables.
}
\source{
\url{https://www.nature.com/articles/s41586-020-1969-6}
}
\description{
This dataset contains less records than \code{data("pcawg_full")} but
with more ID columns. Of note, only white-list donors included.
}
\examples{
\donttest{
load_data("pcawg_simple")
}
}
