% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASETS.R
\docType{data}
\name{tcga}
\alias{tcga}
\title{TCGA Case Identifiers}
\format{
A data frame with 150849 rows and 5 variables.
}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\description{
How to get the dataset can be viewed in code under \code{data-raw}.
Cases in \code{case_id} column can be directly mapped to a GDC portal
page, e.g. \url{https://portal.gdc.cancer.gov/cases/30a1fe5e-5b12-472c-aa86-c2db8167ab23}.
}
\examples{
\donttest{
load_data("tcga")
}
}
