% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_icgc.R
\name{convert_icgc}
\alias{convert_icgc}
\title{Convert ICGC Identifiers}
\usage{
convert_icgc(
  x,
  from = c("icgc_specimen_id", "submitted_specimen_id", "icgc_sample_id",
    "submitted_sample_id", "icgc_donor_id", "submitted_donor_id"),
  to = c("icgc_donor_id", "submitted_donor_id", "icgc_specimen_id",
    "submitted_specimen_id", "icgc_sample_id", "submitted_sample_id")
)
}
\arguments{
\item{x}{A character vector to convert.}

\item{from}{Which identifier type to be converted.}

\item{to}{Identifier type convert to.}
}
\value{
A character vector.
}
\description{
Convert ICGC Identifiers
}
\examples{
convert_icgc("SP29019")
\dontrun{
convert_icgc("SA170678")
}
}
