% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDConverter-package.R
\docType{package}
\name{IDConverter-package}
\alias{IDConverter}
\alias{IDConverter-package}
\title{IDConverter: Convert Identifiers in Biological Databases}
\description{
Identifiers in biological databases connect different levels of metadata, phenotype data or genotype data. This tool is designed to easily convert identifiers within or between different biological databases (Wang, Shixiang, et al. (2021) \doi{10.1371/journal.pgen.1009557}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/IDConverter}
  \item \url{https://wanglabcsu.github.io/IDConverter/}
  \item Report bugs at \url{https://github.com/WangLabCSU/IDConverter/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
