% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data from Local or Remote Zenodo Repository}
\usage{
load_data(x)
}
\arguments{
\item{x}{a dataset name.}
}
\value{
typically a \code{data.frame}, depends on \code{x}.
}
\description{
Data are stored in remote \href{https://zenodo.org/record/6342397}{Zenodo repo}.
This function will help download required data and load it into R.
}
\examples{
\donttest{
load_data("pcawg_full")
load_data("pcawg_simple")
load_data("tcga")
load_data("icgc")
}
}
