% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gdc_uuid.R
\name{parse_gdc_file_uuid}
\alias{parse_gdc_file_uuid}
\title{Parse Metadata from GDC Portal File UUID}
\usage{
parse_gdc_file_uuid(x)
}
\arguments{
\item{x}{a GDC manifest file or a vector of file UUIDs.}
}
\value{
a \code{data.frame}
}
\description{
Parse Metadata from GDC Portal File UUID
}
\examples{
parse_gdc_file_uuid("874e71e0-83dd-4d3e-8014-10141b49f12c")
parse_gdc_file_uuid(
  c(
    "874e71e0-83dd-4d3e-8014-10141b49f12c",
    "2c16506f-1110-4d60-81e3-a85233c79909"
  )
)
}
